+++ src/Evolution/Systems/ForceFree/Subcell/SetInitialRdmpData.cpp
@@ -1,7 +1,7 @@
 // Distributed under the MIT License.
 // See LICENSE.txt for details.
 
-#include "Evolution/Systems/ForceFree/Subcell/InitialDataTci.hpp"
+#include "Evolution/Systems/ForceFree/Subcell/SetInitialRdmpData.hpp"
 
 #include "DataStructures/DataVector.hpp"
 #include "DataStructures/Tensor/EagerMath/Magnitude.hpp"
