+++ tests/Unit/Evolution/Systems/ForceFree/CMakeLists.txt
@@ -10,7 +10,7 @@ set(LIBRARY_SOURCES
   FiniteDifference/Test_Tags.cpp
   Subcell/Test_ComputeFluxes.cpp
   Subcell/Test_GhostData.cpp
-  Subcell/Test_InitialDataTci.cpp
+  Subcell/Test_SetInitialRdmpData.cpp
   Subcell/Test_TciOnDgGrid.cpp
   Subcell/Test_TciOnFdGrid.cpp
   Test_Characteristics.cpp
